services/extractor/pyproject.toml
[project]
name = "egide-extractor"
version = "0.0.1"
description = "Egide document extraction service (Docling + pypdf + python-docx + MarkItDown)"
requires-python = ">=3.13,<3.14"
license = "AGPL-3.0-only"
dependencies = [
    # Web framework
    "fastapi[standard]>=0.115.0",
    "python-multipart>=0.0.20",
    # Primary extractor (cf. CLAUDE.md — Docling primary, no Unstructured)
    "docling>=2.25.0",
    # Fallback extractors
    "pypdf>=5.3.0",
    "python-docx>=1.1.2",
    "markitdown[all]>=0.1.1",
    # Observability
    "structlog>=24.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/extractor"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
